pyproject.toml
[project]
name = "nats-firehose-ingest"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "atproto>=0.0.62",
    "fastapi>=0.118.0",
    "nats-py>=2.11.0",
    "pydantic>=2.11.9",
    "structlog>=25.4.0",
    "uvicorn>=0.37.0",
]
